+++ sparv/modules/xml_export/pretty.py
@@ -142,10 +142,7 @@ def combined(corpus: Corpus = Corpus(),
              version_info: ExportInput = ExportInput("version_info/info_[metadata.id].yaml"),
              include_version_info: bool = Config("xml_export.include_version_info")):
     """Combine XML export files into a single XML file."""
-    if include_version_info:
-        xml_utils.combine(corpus, out, source_files, xml_input, version_info)
-    else:
-        xml_utils.combine(corpus, out, source_files, xml_input)
+    xml_utils.combine(corpus, out, source_files, xml_input, version_info if include_version_info else None)
 
 
 @exporter("Compressed combined XML export", config=[
@@ -156,10 +153,16 @@ def combined(corpus: Corpus = Corpus(),
         datatype=str,
     )
 ])
-def compressed(out: Export = Export("xml_export.combined/[xml_export.filename_compressed]"),
-               xmlfile: ExportInput = ExportInput("xml_export.combined/[xml_export.filename_combined]")):
+def compressed(
+        corpus: Corpus = Corpus(),
+        out: Export = Export("xml_export.combined/[xml_export.filename_compressed]"),
+        source_files: AllSourceFilenames = AllSourceFilenames(),
+        xml_input: ExportInput = ExportInput("xml_export.pretty/[xml_export.filename]", all_files=True),
+        version_info: ExportInput = ExportInput("version_info/info_[metadata.id].yaml"),
+        include_version_info: bool = Config("xml_export.include_version_info")
+) -> None:
     """Compress combined XML export."""
-    xml_utils.compress(xmlfile, out)
+    xml_utils.combine(corpus, out, source_files, xml_input, version_info if include_version_info else None, True)
 
 
 @installer("Copy compressed XML to a target path, optionally on a remote host", config=[